02_rest/requirements.txt
fastapi
pydantic
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
pyjwt
python-multipart
